0.0.8 - Major Over-Haul
-Tools re-implemented differently, less space, cluster, mess, etc
involved.
-Game is now resizable (+ bug)!
-Different methods of tile interaction
-Stone (And other items in future) can be used/manipulated
ToDo:
-Need better rendering system
-Fix how the resize works, because resizing screen physically moves the
character, which is a no-no.
-More Tiles
-Misc

+++ src/Tools/Tool.java
@@ -6,7 +6,7 @@ import java.awt.Rectangle;
 import Core.Game;
 
 public abstract class Tool extends Item {
-	int toolID;
+	byte toolID;
 	public Rectangle toolArea;
 	Game game;
